backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic-settings==2.6.1
paho-mqtt==2.1.0